aplit MintHTTPProvider into 2 modules, 1 for basic HTTP, one for websocket
